backend/requirements.txt
# Core API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
python-dotenv>=1.0.0

# RAG (embeddings + FAISS)
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0

# Development / tests
pytest>=8.0.0
pytest-asyncio>=0.24.0
